+++ src/_packages/remote-client.index.ts
@@ -1,25 +1,25 @@
 // ⚠️ WARNING: This code has been generated so that any manual changes will be overwritten
 // `@promptbook/remote-client`
 
-import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
 import { compilePipelineOnRemoteServer } from '../conversion/compilePipelineOnRemoteServer';
 import { RemoteLlmExecutionTools } from '../llm-providers/remote/RemoteLlmExecutionTools';
 import { preparePipelineOnRemoteServer } from '../prepare/preparePipelineOnRemoteServer';
-import type { PromptbookServer_Identification } from '../remote-server/socket-types/_subtypes/PromptbookServer_Identification';
-import type { PromptbookServer_ApplicationIdentification } from '../remote-server/socket-types/_subtypes/PromptbookServer_Identification';
-import type { PromptbookServer_AnonymousIdentification } from '../remote-server/socket-types/_subtypes/PromptbookServer_Identification';
+import type {
+    AnonymousModeIdentification,
+    ApplicationModeIdentification,
+    Identification,
+} from '../remote-server/socket-types/_subtypes/Identification';
 import type { RemoteClientOptions } from '../remote-server/types/RemoteClientOptions';
-
+import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
 
 // Note: Exporting version from each package
 export { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION };
 
-
 // Note: Entities of the `@promptbook/remote-client`
-export { compilePipelineOnRemoteServer };
-export { RemoteLlmExecutionTools };
-export { preparePipelineOnRemoteServer };
-export type { PromptbookServer_Identification };
-export type { PromptbookServer_ApplicationIdentification };
-export type { PromptbookServer_AnonymousIdentification };
-export type { RemoteClientOptions };
+export { compilePipelineOnRemoteServer, preparePipelineOnRemoteServer, RemoteLlmExecutionTools };
+export type {
+    AnonymousModeIdentification as PromptbookServer_AnonymousIdentification,
+    ApplicationModeIdentification as PromptbookServer_ApplicationIdentification,
+    Identification as PromptbookServer_Identification,
+    RemoteClientOptions,
+};