+++ src/cli/common/$provideLlmToolsForCli.ts
@@ -4,7 +4,7 @@ import { CLI_APP_ID } from '../../config';
 import { UnexpectedError } from '../../errors/UnexpectedError';
 import { $provideLlmToolsForWizzardOrCli } from '../../llm-providers/_common/register/$provideLlmToolsForWizzardOrCli';
 import type { CacheLlmToolsOptions } from '../../llm-providers/_common/utils/cache/CacheLlmToolsOptions';
-import type { ApplicationRemoteServerOptionsLoginResponse } from '../../remote-server/types/RemoteServerOptions';
+import type { LoginResponse } from '../../remote-server/types/RemoteServerOptions';
 import type { string_url } from '../../types/typeAliases';
 import type { really_unknown } from '../../utils/organization/really_unknown';
 import { TODO_USE } from '../../utils/organization/TODO_USE';
@@ -104,7 +104,7 @@ export function $provideLlmToolsForCli(options: ProvideLlmToolsForCliOptions) {
                 });
 
                 const { isSuccess, message, error, identification } =
-                    (await response.json()) as ApplicationRemoteServerOptionsLoginResponse<really_unknown>;
+                    (await response.json()) as LoginResponse<really_unknown>;
 
                 TODO_USE(error);